Ok, I almost found the bug. Just a bit more debugging...
Also, use long for getVolume() (volume can get quite big).

+++ src/geometry/Parallelepiped.java
@@ -97,7 +97,7 @@ public final class Parallelepiped {
         }
     }
 
-    public int getVolume() {
+    public long getVolume() {
         int volumeSoFar = 1;
 
         for (Axis axis : Axis.values()) {
